[Docs] Data 폴더로 관련 파일 이동
