+++ MarketPlace/Domain/Entities/Types/CouponType.swift
@@ -16,11 +16,4 @@ enum CouponType: String, CaseIterable {
         case .refundableCoupon: return "PAYBACK"
         }
     }
-    
-    func toUIName() -> String {
-        switch self {
-        case .giftableCoupon: return "증정형 쿠폰"
-        case .refundableCoupon: return "환급형 쿠폰"
-        }
-    }
 }
